fix 64 bit issue
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@23486 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/iptables/include/linux/netfilter_ipv4/ipt_MARK.h
@@ -25,7 +25,7 @@ struct ipt_mark_target_info_v1 {
 };
 
 struct xt_mark_tginfo2 {
-	unsigned long mark, mask;
+	u_int32_t mark, mask;
 };
 
 #endif /*_IPT_MARK_H_target*/
